+++ src/file-history.h
@@ -3,9 +3,9 @@
 
 #include <stdbool.h>
 
-void add_file_history(int row, int col, const char *filename);
+void add_file_history(unsigned long row, unsigned long col, const char *filename);
 void load_file_history(const char *filename);
 void save_file_history(const char *filename);
-bool find_file_in_history(const char *filename, int *row, int *col);
+bool find_file_in_history(const char *filename, unsigned long *row, unsigned long *col);
 
 #endif
